refactor: move localStorage.js to app/data
localStorage.js is a plain persistence helper module, not a Fluxible
store, so it belongs alongside FavouriteData.js in app/data rather
than app/store. Updated all importers (Fluxible stores, components,
utils, tests) to the new path.
