+++ app/component/itinerary/customizesearch/PersonalizeAgainModal.js
@@ -5,7 +5,7 @@ import { RadioGroup } from '@hsl-fi/form';
 import { Text } from '@hsl-fi/layout-primitives';
 import { useIntl } from 'react-intl';
 import { useConfigContext } from '../../../configurations/ConfigContext';
-import { setPersonalization } from '../../../store/localStorage';
+import { setPersonalization } from '../../../data/localStorage';
 
 export default function PersonalizeAgainModal({ open, onClose, onContinue }) {
   const intl = useIntl();
