+++ app/component/nearyou/CityBikeInfo.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { getReadMessageIds, setReadMessageId } from '../../store/localStorage';
+import { getReadMessageIds, setReadMessageId } from '../../data/localStorage';
 import { useConfigContext } from '../../configurations/ConfigContext';
 import {
   getRentalNetworkConfig,
