+++ app/store/OldSearchesStore.js
@@ -5,7 +5,10 @@ import isEqual from 'lodash/isEqual';
 import orderBy from 'lodash/orderBy';
 import { getNameLabel } from '@digitransit-search-util/digitransit-search-util-uniq-by-label';
 import { unixTime } from '../util/timeUtils';
-import { getOldSearchesStorage, setOldSearchesStorage } from './localStorage';
+import {
+  getOldSearchesStorage,
+  setOldSearchesStorage,
+} from '../data/localStorage';
 
 /**
  * The current version number of this store.
