+++ app/store/PositionStore.js
@@ -1,5 +1,5 @@
 import Store from 'fluxible/addons/BaseStore';
-import { getGeolocationState, setGeolocationState } from './localStorage';
+import { getGeolocationState, setGeolocationState } from '../data/localStorage';
 
 export default class PositionStore extends Store {
   static storeName = 'PositionStore';
