+++ test/unit/component/MessageBar.test.js
@@ -7,7 +7,7 @@ import {
 } from '../../../app/component/MessageBar';
 import { mockContext } from '../helpers/mock-context';
 import { renderWithProviders } from '../helpers/mock-providers';
-import { setReadMessageIds } from '../../../app/store/localStorage';
+import { setReadMessageIds } from '../../../app/data/localStorage';
 import { AlertSeverityLevelType } from '../../../app/constants';
 
 const defaultProps = {
